Move SSE client logic to model class + RemoteSR
